Refactor basic/test rebasers into their own file
