+++ experimental/dds/tree2/src/feature-libraries/defaultFieldKinds.ts
@@ -74,61 +74,6 @@ function brandedFieldKind<
 }
 
 /**
- * Picks the last value written.
- *
- * TODO: it seems impossible for this to obey the desired axioms.
- * Specifically inverse needs to cancel, restoring the value from the previous change which was discarded.
- */
-export function lastWriteWinsRebaser<TChange>(data: {
-	noop: TChange;
-	invert: (changes: TChange) => TChange;
-}): FieldChangeRebaser<TChange> {
-	const compose = (changes: TChange[]) =>
-		changes.length >= 0 ? changes[changes.length - 1] : data.noop;
-	const rebase = (change: TChange, _over: TChange) => change;
-	return referenceFreeFieldChangeRebaser({ ...data, compose, rebase });
-}
-
-export interface Replacement<T> {
-	old: T;
-	new: T;
-}
-
-export type ReplaceOp<T> = Replacement<T> | 0;
-
-/**
- * Picks the last value written.
- *
- * Consistent if used on valid paths with correct old states.
- */
-export function replaceRebaser<T>(): FieldChangeRebaser<ReplaceOp<T>> {
-	return referenceFreeFieldChangeRebaser({
-		rebase: (change: ReplaceOp<T>, over: ReplaceOp<T>) => {
-			if (change === 0) {
-				return 0;
-			}
-			if (over === 0) {
-				return change;
-			}
-			return { old: over.new, new: change.new };
-		},
-		compose: (changes: ReplaceOp<T>[]) => {
-			const f = changes.filter((c): c is Replacement<T> => c !== 0);
-			if (f.length === 0) {
-				return 0;
-			}
-			for (let index = 1; index < f.length; index++) {
-				assert(f[index - 1].new === f[index].old, 0x3a4 /* adjacent replaces must match */);
-			}
-			return { old: f[0].old, new: f[f.length - 1].new };
-		},
-		invert: (changes: ReplaceOp<T>) => {
-			return changes === 0 ? 0 : { old: changes.new, new: changes.old };
-		},
-	});
-}
-
-/**
  * ChangeHandler that only handles no-op / identity changes.
  */
 export const noChangeHandler: FieldChangeHandler<0> = {
