+++ experimental/dds/tree2/src/test/feature-libraries/modular-schema/genericFieldKind.spec.ts
@@ -4,11 +4,7 @@
  */
 
 import { strict as assert } from "assert";
-import { TUnsafe, Type } from "@sinclair/typebox";
 import {
-	FieldChangeHandler,
-	FieldKind,
-	Multiplicity,
 	NodeChangeset,
 	GenericChangeset,
 	genericFieldKind,
@@ -16,35 +12,13 @@ import {
 	CrossFieldManager,
 	RevisionMetadataSource,
 } from "../../../feature-libraries";
-// TODO: this is not the file being tested, importing it should not be required here.
-// eslint-disable-next-line import/no-internal-modules
-import * as FieldKinds from "../../../feature-libraries/defaultFieldKinds";
 import { makeAnonChange, tagChange, TaggedChange, Delta, FieldKey } from "../../../core";
-import { brand, fail } from "../../../util";
+import { brand } from "../../../util";
 import { fakeTaggedRepair as fakeRepair, makeEncodingTestSuite } from "../../utils";
-import { IJsonCodec, makeCodecFamily, makeValueCodec } from "../../../codec";
+import { IJsonCodec } from "../../../codec";
 import { singleJsonCursor } from "../../../domains";
+import { ValueChangeset, valueField, valueHandler } from "./basicRebasers";
 
-type ValueChangeset = FieldKinds.ReplaceOp<number>;
-
-const valueHandler: FieldChangeHandler<ValueChangeset> = {
-	rebaser: FieldKinds.replaceRebaser(),
-	codecsFactory: () =>
-		makeCodecFamily([[0, makeValueCodec<TUnsafe<ValueChangeset>>(Type.Any())]]),
-	editor: { buildChildChange: () => fail("Child changes not supported") },
-	intoDelta: (change) =>
-		change === 0 ? [] : [{ type: Delta.MarkType.Modify, setValue: change.new }],
-
-	isEmpty: (change) => change === 0,
-};
-
-const valueField = new FieldKind(
-	brand("Value"),
-	Multiplicity.Value,
-	valueHandler,
-	() => false,
-	new Set(),
-);
 const valueFieldKey: FieldKey = brand("Value");
 
 const fieldA: FieldKey = brand("a");
